% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCriticalDifference.R
\name{criticalDifferenceDiagram}
\alias{criticalDifferenceDiagram}
\title{Critical Difference Diagram}
\usage{
criticalDifferenceDiagram(
  data,
  method = "NemenyiCD",
  alpha = 0.05,
  cex = 0.75,
  ...
)
}
\arguments{
\item{data}{Data Frame or Matrix with results of each algorithm / scenario}

\item{method}{Method by which to perform comparisons}

\item{alpha}{Significance level to get the critical difference.}

\item{cex}{Font Size}

\item{...}{Additional arguments for \code{\link{rankMatrix}}}
}
\value{
Plot (TODO: silent object with pairs)
}
\description{
Creates a critical difference diagram as shown in Demsar (2006)
}
\examples{
# Not Run
# criticalDifferenceDiagram(someData)
}
\references{
Demsar, J. (2006) Statistical Comparisons of Classifiers over Multiple Data Sets.
}
