% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi.R
\name{nemenyiCDVal}
\alias{nemenyiCDVal}
\title{Calculate Nemenyi Critical Difference Value}
\usage{
nemenyiCDVal(alpha = 0.05, num.classifiers, num.scenarios)
}
\arguments{
\item{alpha}{The significance size. E.g 95\% is 0.05}

\item{num.classifiers}{Number of classifiers tested}

\item{num.scenarios}{Number of scenarios of implementation}
}
\value{
Critical Difference Value
}
\description{

}
\details{
Not to be mistaken with the Nemenyi Test
}
\examples{
cd = nemenyiCDVal(num.classifiers = 3, num.scenarios = 4)
}
\references{
Nemenyi, P.B. (1963) Distribution-free Multiple Comparisons. PhD thesis, Princeton University.

Demsar, J. (2006) Statistical Comparisons of Classifiers over Multiple Data Sets.
}
