% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi.R
\name{nemenyiTest}
\alias{nemenyiTest}
\title{Nemenyi Test
The Nemenyi test is a post-hoc test intended to find the groups of data that
differ aftera  statistical test of multiple comparisons (such as the
Friedman test) has rejected the null hypothesis that the performance of the
comparisons on the groups of data is similar.}
\usage{
nemenyiTest()
}
\description{
Nemenyi Test
The Nemenyi test is a post-hoc test intended to find the groups of data that
differ aftera  statistical test of multiple comparisons (such as the
Friedman test) has rejected the null hypothesis that the performance of the
comparisons on the groups of data is similar.
}
